requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=1.4.0,<2.0.0
pymysql==1.1.0
cryptography>=41.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
email-validator==2.1.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
aiosmtplib>=3.0.0
jinja2>=3.1.0
